+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Notification when a new version of Wombat is available to download.
+- Notification when a new version of Wombat is available to download
+- Service and Method select dropdowns are searchable
 
 ### Fixed
 - Support mac case-sensitive filesystem. Thanks to [@Azuka](https://github.com/Azuka)