+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- custom rendering/parsing support for Google Well Known [wrapper] Types
-- Notification when a new version of Wombat is available to download.
+- Custom rendering/parsing support for Google Well Known [wrapper] Types
+- Notification when a new version of Wombat is available to download
+- Service and Method select dropdowns are searchable
 
 ### Fixed
 - Support mac case-sensitive filesystem. Thanks to [@Azuka](https://github.com/Azuka)
+- UI unresponsive when rendering empty state for repeated fields
 
 ## [v0.2.2] - 2020-11-12